requirements.txt
python-dotenv==0.21.0
aiogram==2.22.2
requests==2.25.1
folium==0.13.0
google==3.0.0
googletrans==3.1.0a0
youtube-search-python==1.6.6
moviepy==1.0.3
pytube==12.1.2